restructured project into modules, simulator jar with dependencies, webapp jar
