--- src/test/java/SE459/CleanSweep/CleanSweepApplicationTests.java
@@ -1,13 +1,0 @@
-package SE459.CleanSweep;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CleanSweepApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
