+++ webapp/src/test/java/SE459/CleanSweep/webapp/controller/CleanSweepApplicationTests.java
@@ -1,0 +1,13 @@
+package SE459.CleanSweep.webapp.controller;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CleanSweepApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
